readme更新，修复编辑用户select选中问题

+++ release-upload-pull.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ x"$1" = x ]; then
+    echo "请输入分支参数"
+    exit 1
+fi
+
 rm zhuque
 echo "上传文件......"
 rz
